+++ tsup.config.ts
@@ -1,7 +1,11 @@
 import { defineConfig } from 'tsup'
 
 export default defineConfig({
-  entry: ['src/index.ts'],
+  entry: [
+    'src/index.ts',
+    'src/integrations/vitest.ts',
+    'src/integrations/jest.ts',
+  ],
   format: ['esm', 'cjs'],
   dts: true,
   clean: true,
@@ -9,4 +13,5 @@ export default defineConfig({
   splitting: false,
   treeshake: true,
   target: 'node18',
+  external: ['vitest', '@jest/globals'],
 })